+ processor + changed datatype in domain model and dto from int to double + added regexp validation for double values + fixed navigation previous page error in preprocessorui

+++ src/main/java/ru/mikaev/sapr/domain/Knot.java
@@ -17,5 +17,5 @@ public class Knot {
     @GeneratedValue
     private Long id;
 
-    private int load;
+    private double load;
 }
